refactor: Move utils.search_engine.entity_base to utils.common.entities
