+++ utils/search_engine/entity.py
@@ -2,7 +2,7 @@ from typing import Final
 
 import polars as pl
 
-from utils.search_engine.entity_base import Entity, SubEntity, field
+from utils.common.entities import Entity, SubEntity, field
 
 
 class Stats(Entity):
